requirements.txt
typer==0.12.3
fastapi==0.112.2
uvicorn[standard]==0.30.6
streamlit==1.37.0
sqlmodel==0.0.22
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
loguru==0.7.2
pydantic==2.9.2
boto3==1.35.23
botocore==1.35.23
s3fs==2024.6.0
awscli==1.34.23
neptune-python-utils==1.24.0
numpy==2.1.1
pandas==2.2.3
scikit-learn==1.5.1
scipy==1.14.1
networkx==3.3
rdflib==7.1.0
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
neo4j>=5.22.0
transformers==4.43.3
accelerate==0.33.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
torch-geometric==2.5.3
pyg-lib==0.4.0
peft==0.11.1
datasets==3.0.1
evaluate==0.4.2
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
ruff==0.6.8
pytest==8.3.2

